Move gcheap walk helpers to opt-in
