+++ src/coreclr/nativeaot/Runtime/eventtrace_gcheap.cpp
@@ -22,6 +22,7 @@
 
 #include "eventtrace_etw.h"
 #include "eventtracepriv.h"
+#include "profheapwalkhelper.h"
 
 /****************************************************************************/
 /* Methods that are called from the runtime */
@@ -963,4 +964,132 @@ void ETW::GCLog::EndHeapDump(ProfilerWalkHeapContext* profilerWalkHeapContext)
     // Delete any GC state built up in the context
     profilerWalkHeapContext->pvEtwContext = NULL;
     delete pContext;
+}
+
+namespace
+{
+    void ProfScanRootsHelper(Object** ppObject, ScanContext* pSC, uint32_t dwFlags)
+    {
+        Object* pObj = *ppObject;
+        if (dwFlags& GC_CALL_INTERIOR)
+        {
+            pObj = GCHeapUtilities::GetGCHeap()->GetContainingObject(pObj, true);
+            if (pObj == nullptr)
+                return;
+        }
+        ScanRootsHelper(pObj, ppObject, pSC, dwFlags);
+    }
+
+    void GcScanRootsForETW(promote_func* fn, int condemned, int max_gen, ScanContext* sc)
+    {
+        UNREFERENCED_PARAMETER(condemned);
+        UNREFERENCED_PARAMETER(max_gen);
+
+        FOREACH_THREAD(pThread)
+        {
+            if (pThread->IsGCSpecial())
+                continue;
+
+            if (GCHeapUtilities::GetGCHeap()->IsThreadUsingAllocationContextHeap(pThread->GetAllocContext(), sc->thread_number))
+                continue;
+
+            sc->thread_under_crawl = pThread;
+            sc->dwEtwRootKind = kEtwGCRootKindStack;
+            pThread->GcScanRoots(reinterpret_cast<void*>(fn), sc);
+            sc->dwEtwRootKind = kEtwGCRootKindOther;
+        }
+        END_FOREACH_THREAD
+    }
+
+    void ScanHandleForETW(Object** pRef, Object* pSec, uint32_t flags, ScanContext* context, bool isDependent)
+    {
+        ProfilingScanContext* pSC = (ProfilingScanContext*)context;
+
+        // Notify ETW of the handle
+        if (ETW::GCLog::ShouldWalkHeapRootsForEtw())
+        {
+            ETW::GCLog::RootReference(
+                pRef,
+                *pRef,          // object being rooted
+                pSec,           // pSecondaryNodeForDependentHandle
+                isDependent,
+                pSC,
+                0,              // dwGCFlags,
+                flags);     // ETW handle flags
+        }
+    }
+
+    // This is called only if we've determined that either:
+    //     a) The Profiling API wants to do a walk of the heap, and it has pinned the
+    //     profiler in place (so it cannot be detached), and it's thus safe to call into the
+    //     profiler, OR
+    //     b) ETW infrastructure wants to do a walk of the heap either to log roots,
+    //     objects, or both.
+    // This can also be called to do a single walk for BOTH a) and b) simultaneously.  Since
+    // ETW can ask for roots, but not objects
+    void GCProfileWalkHeapWorker(BOOL fShouldWalkHeapRootsForEtw, BOOL fShouldWalkHeapObjectsForEtw)
+    {
+        ProfilingScanContext SC(FALSE);
+        unsigned max_generation = GCHeapUtilities::GetGCHeap()->GetMaxGeneration();
+
+        // **** Scan roots:  Only scan roots if profiling API wants them or ETW wants them.
+        if (fShouldWalkHeapRootsForEtw)
+        {
+            GcScanRootsForETW(&ProfScanRootsHelper, max_generation, max_generation, &SC);
+            SC.dwEtwRootKind = kEtwGCRootKindFinalizer;
+            GCHeapUtilities::GetGCHeap()->DiagScanFinalizeQueue(&ProfScanRootsHelper, &SC);
+
+            // Handles are kept independent of wks/svr/concurrent builds
+            SC.dwEtwRootKind = kEtwGCRootKindHandle;
+            GCHeapUtilities::GetGCHeap()->DiagScanHandles(&ScanHandleForETW, max_generation, &SC);
+        }
+
+        // **** Scan dependent handles: only if ETW wants roots
+        if (fShouldWalkHeapRootsForEtw)
+        {
+            // GcScanDependentHandlesForProfiler double-checks
+            // CORProfilerTrackConditionalWeakTableElements() before calling into the profiler
+
+            ProfilingScanContext* pSC = &SC;
+
+            // we'll re-use pHeapId (which was either unused (0) or freed by EndRootReferences2
+            // (-1)), so reset it to NULL
+            _ASSERTE((*((size_t *)(&pSC->pHeapId)) == (size_t)(-1)) ||
+                    (*((size_t *)(&pSC->pHeapId)) == (size_t)(0)));
+            pSC->pHeapId = NULL;
+
+            GCHeapUtilities::GetGCHeap()->DiagScanDependentHandles(&ScanHandleForETW, max_generation, &SC);
+        }
+
+        ProfilerWalkHeapContext profilerWalkHeapContext(FALSE, SC.pvEtwContext);
+
+        // **** Walk objects on heap: only if ETW wants them.
+        if (fShouldWalkHeapObjectsForEtw)
+        {
+            GCHeapUtilities::GetGCHeap()->DiagWalkHeap(&HeapWalkHelper, &profilerWalkHeapContext, max_generation, true /* walk the large object heap */);
+        }
+
+        // **** Done! Indicate to ETW helpers that the heap walk is done, so any buffers
+        // should be flushed into the ETW stream
+        if (fShouldWalkHeapObjectsForEtw || fShouldWalkHeapRootsForEtw)
+        {
+            ETW::GCLog::EndHeapDump(&profilerWalkHeapContext);
+        }
+    }
+}
+
+void ETW::GCLog::WalkHeap()
+{
+    if (ETW::GCLog::ShouldWalkStaticsAndCOMForEtw())
+        ETW::GCLog::WalkStaticsAndCOMForETW();
+
+    BOOL fShouldWalkHeapRootsForEtw = ETW::GCLog::ShouldWalkHeapRootsForEtw();
+    BOOL fShouldWalkHeapObjectsForEtw = ETW::GCLog::ShouldWalkHeapObjectsForEtw();
+
+    // we need to walk the heap if one of GC_PROFILING or FEATURE_EVENT_TRACE
+    // is defined, since both of them make use of the walk heap worker.
+    if (fShouldWalkHeapRootsForEtw || fShouldWalkHeapObjectsForEtw)
+    {
+        GCProfileWalkHeapWorker(fShouldWalkHeapRootsForEtw, fShouldWalkHeapObjectsForEtw);
+    }
 }
